refactor: more lua rework
