+++ nvim/lua/plugin/treesitter.lua
@@ -21,6 +21,11 @@ return function()
     },
     highlight = {
       enable = true
-    }
+    },
+    indent = { enable = true },
+    autopairs = { enable = true },
+    rainbow = { enable = true },
+    autotag = { enable = true },
+    context_commentstring = { enable = true },
   }
 end
